completed online eyelink event handling + more
- eyelink event handling is now almost complete, button and message
events are remaining.
- added ability to retrieve events for a specific device by access the
client side device object's getEvent method directly.
- added a getEvents and clearEvents convienience method to the
simpleExperimentRuntime class. These can be used to easily get or clear
the global device event queue or a specific devices event deque if the
device's label is provided. User can also indicate if the retrieved
events should be in list, dict, or ioHubEvent Object format (each
increases the time to get the events; getting the events as dict's does
not seems to add any significant time to baseline, so it may be the way
to go by default, we will see.
- other misc. fixed.
- sorting the global event list by time was causing a random int out of
bounds error in python ( apparently a known bug in python 2.x), so that
is disabled for now and a bug is raised in github to look into how to
fix it.