OSX Unicode key event support
The commit contains updated code for the OSX unicode key event support.
Seems to be working OK on my mac when I use the Option key and pres
another key to generate unicode symbols. Testing on a non US keyboard
layout will be important.

The 'Simple' example can be used for testing the unicode support and
viewing the unicode characters in a psychopy window that is using the
text stim. See the comments in the simple examples run.py for more
details.