finished log file integration testing
Added two examples:
- psychopyIntegrationTest/run.py: shows the logging and time base
integration as it stands now. Also has some updates to the
iohubQuickConnection function to make it more useful for a wider set of
applications (so other script only examples may be broken right now).
The example also tests that the integration is working as expected to a
degree.
- dataStore/saveDataStoreLog.py: Example of how to select a dataStore
HDF5 and save out the events for a given iohub device event type along
with the per session data. In the example, the experiment.LogEvent event
type is used, so the example saves out the experiment log entries that
the iohub either generated or received from the experiment.